updated for v0.4

+++ lib/ext/float_ext.rb
@@ -1,5 +1,5 @@
 require 'number_to_cn/methods.rb'
 
-class Integer
+class Float
   include NumberToCn
 end
